backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
python-socketio==5.11.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
aiosqlite==0.19.0

# AI/ML APIs
google-generativeai==0.3.2
openai==1.10.0
anthropic==0.18.1

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Notifications
twilio==8.12.0

# Data Processing
numpy==1.26.3
pandas==2.2.0
scipy==1.12.0

# Configuration
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Validation
pytest>=7.4.0,<8.0.0
pytest-asyncio>=0.21.0
pytest-cov==4.1.0

# Utilities
python-multipart==0.0.6
email-validator==2.1.0
jinja2==3.1.3
